Trait fadroma_storage::traits1::Storable

pub trait Storable: Serialize + DeserializeOwned {
    fn key(&self) -> StdResult<Vec<u8>>;

    fn save<S: Storage, A: Api, Q: Querier>(
        &self, 
        deps: &mut Extern<S, A, Q>
    ) -> StdResult<()> { ... }

    fn remove<S: Storage, A: Api, Q: Querier>(
        self, 
        deps: &mut Extern<S, A, Q>
    ) -> StdResult<()> { ... }

    fn namespace() -> Vec<u8> { ... }

    fn concat_key(key: &[u8]) -> Vec<u8> { ... }

    fn load<S: Storage, A: Api, Q: Querier>(
        deps: &Extern<S, A, Q>, 
        key: &[u8]
    ) -> StdResult<Option<Self>> { ... }

    fn static_save<S: Storage, A: Api, Q: Querier>(
        deps: &mut Extern<S, A, Q>, 
        key: &[u8], 
        item: &Self
    ) -> StdResult<()> { ... }

    fn static_remove<S: Storage, A: Api, Q: Querier>(
        deps: &mut Extern<S, A, Q>, 
        key: &[u8]
    ) -> StdResult<()> { ... }
}

Trait that will add storage options to your struct, at minimum, you will have to implement key() method that will let storage know where to save or from where to load your data.

Optionally, you can define the namespace of your struct that will namespace it when storing.

use serde::{Deserialize, Serialize};

// use fadroma::{platform::*, storage::traits1::*};
use fadroma_platform_scrt::*;
use fadroma_storage::traits1::*;
 
#[derive(Deserialize, Serialize)]
struct Config {
    some_value: String,
}
 
impl Storable for Config {
    fn key(&self) -> StdResult<Vec<u8>> {
        Ok(to_vec(b"some_key")?)
    }
}
 
fn init<S: Storage, A: Api, Q: Querier>(deps: &mut Extern<S, A, Q>) -> StdResult<()> {
    let config = Config { some_value: "foo".to_string() };
 
    config.save(deps)?;
 
    let maybe_config: Option<Config> = Config::load(&deps, b"some_key")?;
 
    Ok(())
}

Required methods

fn key(&self) -> StdResult<Vec<u8>>

Storage key used for saving Self

Provided methods

fn save<S: Storage, A: Api, Q: Querier>(
&self,
deps: &mut Extern<S, A, Q>
) -> StdResult<()>

Save Self in the storage

fn remove<S: Storage, A: Api, Q: Querier>(
self,
deps: &mut Extern<S, A, Q>
) -> StdResult<()>

Remove Self from storage

fn namespace() -> Vec<u8>

Static: namespace for self

fn concat_key(key: &[u8]) -> Vec<u8>

Static: Concat of namespace and key

fn load<S: Storage, A: Api, Q: Querier>(
deps: &Extern<S, A, Q>,
key: &[u8]
) -> StdResult<Option<Self>>

Static: Load Self from the storage

fn static_save<S: Storage, A: Api, Q: Querier>(
deps: &mut Extern<S, A, Q>,
key: &[u8],
item: &Self
) -> StdResult<()>

Static: Save Self in the storage

fn static_remove<S: Storage, A: Api, Q: Querier>(
deps: &mut Extern<S, A, Q>,
key: &[u8]
) -> StdResult<()>

Static: Remove Self from storage